+++ CHANGELOG.md
@@ -14,8 +14,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- Broadcaster can use `!theme follower` or `!theme !follower` to activate or deactivate follower only mode
 - Added tests for Themer class
+- Broadcaster can use `!theme follower only` or `!theme follower only off` to activate or deactivate follower only mode
+- Broadcaster can use `!theme sub only` or `!theme sub only off` to activate or deactivate sub only mode
 
 
 ## [0.0.6] - 2019-05-12